requirements.txt
streamlit
torch
torchvision
Pillow
openai
streamlit_option_menu
matplotlib
openpyxl
seaborn